add optional types to metrics
The new prometheus format has types. Some scrapers (like datadog agent)
use the types to map from prometheus types to types in their service.

This is the spec which includes information about types here.
https://github.com/prometheus/docs/blob/master/content/docs/instrumenting/exposition_formats.md

This are the datadog docs which show how the types are mapped. This
shows that there's a major use-case for types.
https://docs.datadoghq.com/integrations/guide/prometheus-metrics/

The approach to add type support was to add additional versions of
exported functions that write types. This is an additive change which
does not break the package API.