support Prometheus compatible histograms (#93)
as a follow-up/alternative to #45 and to solve #28, I'm starting this PR
which adds support for [OTLP
histograms](https://opentelemetry.io/docs/specs/otel/metrics/data-model/#histogram)
(the common `le` - static bucket type of histograms, not the complex
exponential histograms/native histograms).

I know that VM focus is on log based buckets and really appreciate
those, but having the opportunity to use "normal" histograms with this
library would be greatly appreciated, as documented in #28

before working too much on this, I'd like to know whether support for
normal / OTLP style buckets could be considered or not. I don't want to
spend too much time on this PR if at the end @valyala @hagen1778 you
decide against such a feature.

I'd also like to start a discussion on the name. I picked
`OTLPHistogram` in reference to the histograms described in the
OpenTelemetry Protocol, but perhaps the naming is bad.
[EDIT]: I think that `CompatibleHistogram` would make more sense than
`OTLPHistogram`, had not thought about it until after I started
implementing the functions


currently the basic functionality is implemented, and a testcase covers
a simple usage. if there is a decision in favor of this PR, I will add
the missing testcases and complete the implementation (add a function
for custom buckets/upper bounds for example)

---------

Co-authored-by: Roman Khavronenko <hagen1778@gmail.com>
Co-authored-by: hagen1778 <roman@victoriametrics.com>