change the order of summary and quantile(s) (#105)
This is a follow-up to
https://github.com/VictoriaMetrics/metrics/pull/99.

Prior to #99, the order of summary metrics (`_sum` and `_count`) and
quantiles could vary depending on the metric name.

After #99, the order was fixed to show summaries first, followed by
quantiles. This ordering was based on the author's tests, but at the
time, the author was unaware that different orders were possible.

In conclusion:
1. The order should be fixed, which was accomplished in the previous
pull request.
2. The fixed order should match Prometheus or the most common ordering
from versions before #99.
    - The current version (v1.40.1) fixed this order in an unusual way.

Therefore, this pull request modifies the sorting logic for summaries
and quantiles, enforcing that quantiles are printed first, followed by
summaries.

---------

Co-authored-by: Max Kotliar <mkotlyar@victoriametrics.com>